requirements.txt
customtkinter==5.2.2
pillow==10.1.0
svgwrite==1.4.3
cairosvg==2.7.1
rembg==2.0.60
PyInstaller==6.4.0
tkinterdnd2==0.3.0
